move rustc::lint::{context, passes} to rustc_lint.
Also do some cleanup of the interface.
